bernd: Added trusted room concept and room invite accepting
The bot now automatically accepts room invites. Trusted rooms, which
should be joined right away, can be specified in the config.ini. For
this, exclusively internal room IDs (!longroomid:server.org) need to be
used.

Furthermore every plugin is provided with a TRUSTED_ROOMS environment
variable, which contains a list of all trusted room internal IDs in
string representation. This enables a developer to restrict plugin
functionality to certain rooms.